DIGEQ-94 removing manage.py and putting the code inside srunner, this should remove the docker timing issue
